move attribution component
